+++ Westwind.Globalization.Test/Properties/Resources.cs
@@ -1,0 +1,111 @@
+using System;
+using System.Web;
+using Westwind.Globalization;
+
+namespace Westwind.Globalization.Test
+{
+    public class GeneratedResourceSettings
+    {
+        // You can change the ResourceAccess Mode globally in Application_Start        
+        public static ResourceAccessMode ResourceAccessMode = ResourceAccessMode.AspNetResourceProvider;
+    }
+
+	public class CommonWords
+	{
+		public static System.String Save
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("CommonWords","Save");
+				return DbRes.T("Save","CommonWords");
+			}
+		}
+
+		public static System.String Cancel
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("CommonWords","Cancel");
+				return DbRes.T("Cancel","CommonWords");
+			}
+		}
+
+		public static System.String Edit
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("CommonWords","Edit");
+				return DbRes.T("Edit","CommonWords");
+			}
+		}
+
+	}
+
+	public class Resources
+	{
+		public static System.String NameIsRequired
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","NameIsRequired");
+				return DbRes.T("NameIsRequired","Resources");
+			}
+		}
+
+		public static System.String Today
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","Today");
+				return DbRes.T("Today","Resources");
+			}
+		}
+
+		public static System.String Yesterday
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","Yesterday");
+				return DbRes.T("Yesterday","Resources");
+			}
+		}
+
+		public static System.String HelloWorld
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","HelloWorld");
+				return DbRes.T("HelloWorld","Resources");
+			}
+		}
+
+		public static System.String Ready
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","Ready");
+				return DbRes.T("Ready","Resources");
+			}
+		}
+
+		public static System.String AddressIsRequired
+		{
+			get
+			{
+				if (GeneratedResourceSettings.ResourceAccessMode == ResourceAccessMode.AspNetResourceProvider)
+					return (System.String) HttpContext.GetGlobalResourceObject("Resources","AddressIsRequired");
+				return DbRes.T("AddressIsRequired","Resources");
+			}
+		}
+
+	}
+
+}